moves fileupload into separate file structure
